[client/server] refactor code NowCredentials -> MediaCredentials and separete controllers.
