+++ client/src/stores/SettingStore.tsx
@@ -1,6 +1,6 @@
 import { observable, action } from 'mobx';
-import { NowController } from '@openapi/routes'
-import { NowCredentials } from '@openapi/schemas'
+import { MediaController } from '@openapi/routes'
+import { MediaCredentials } from '@openapi/schemas'
 
 
 export class SettingStore {
@@ -8,13 +8,13 @@ export class SettingStore {
     public static scope_playback: string = 'user-read-playback-state';
     public static scope_playlist: string = 'playlist-modify-public';
 
-    @observable playbackCredential: NowCredentials | undefined = undefined;
-    @observable playlistCredential: NowCredentials | undefined = undefined;
+    @observable playbackCredential: MediaCredentials | undefined = undefined;
+    @observable playlistCredential: MediaCredentials | undefined = undefined;
 
     @action
     obtainPlaybackCredential = async () => {
         try {
-            let data: NowCredentials[] = await NowController.find(SettingStore.scope_playback);
+            let data: MediaCredentials[] = await MediaController.find(SettingStore.scope_playback);
             if (data.length > 0)
                 this.playbackCredential = data[0];
         } catch (error) {
@@ -25,7 +25,7 @@ export class SettingStore {
     @action
     obtainPlaylistCredential = async () => {
         try {
-            let data: NowCredentials[] = await NowController.find(SettingStore.scope_playlist);
+            let data: MediaCredentials[] = await MediaController.find(SettingStore.scope_playlist);
             if (data.length > 0)
                 this.playlistCredential = data[0];
         } catch (error) {
@@ -38,7 +38,7 @@ export class SettingStore {
         try {
             if (this.playbackCredential !== undefined) {
                 if (this.playbackCredential.id !== undefined)
-                    await NowController.deleteById(this.playbackCredential.id)
+                    await MediaController.deleteById(this.playbackCredential.id)
                 this.playbackCredential = undefined;
             }
         } catch (error) {
@@ -51,7 +51,7 @@ export class SettingStore {
         try {
             if (this.playlistCredential !== undefined) {
                 if (this.playlistCredential.id !== undefined)
-                    await NowController.deleteById(this.playlistCredential.id)
+                    await MediaController.deleteById(this.playlistCredential.id)
                 this.playlistCredential = undefined;
             }
         } catch (error) {
