+++ server/src/models/index.ts
@@ -1,4 +1,4 @@
-export * from './now-credentials.model';
+export * from './media-credentials.model';
 export * from './persistent-key.model';
 export * from './song.model'
 export * from './user.model';
