+++ server/src/repositories/user.repository.ts
@@ -5,9 +5,10 @@ import {
   HasManyRepositoryFactory,
 } from '@loopback/repository';
 import { MongoDataSource } from '../datasources';
-import { User, UserCredentials, Song } from '../models';
+import { User, UserCredentials, Song, MediaCredentials } from '../models';
 import { inject, Getter } from '@loopback/core';
 import { UserCredentialsRepository, SongRepository } from '../repositories';
+import { MediaCredentialsRepository } from './media-credentials.repository';
 
 export type Credentials = {
   email: string;
@@ -18,11 +19,14 @@ export class UserRepository extends DefaultCrudRepository<User, typeof User.prot
 
   public readonly userCredentials: HasOneRepositoryFactory<UserCredentials, typeof User.prototype.id>;
   public readonly songs: HasManyRepositoryFactory<Song, typeof Song.prototype.id>;
+  public readonly mediaCredentials: HasManyRepositoryFactory<MediaCredentials, typeof MediaCredentials.prototype.id>;
 
   constructor(
     @inject('datasources.mongo') protected datasource: MongoDataSource,
     @repository.getter('UserCredentialsRepository') protected userCredentialsRepositoryGetter: Getter<UserCredentialsRepository>,
-    @repository.getter('SongRepository') protected songRepositoryGetter: Getter<SongRepository>) {
+    @repository.getter('SongRepository') protected songRepositoryGetter: Getter<SongRepository>,
+    @repository.getter('MediaCredentialsRepository') protected mediaCredentialsRepositoryGetter: Getter<MediaCredentialsRepository>
+  ) {
     super(User, datasource);
     this.userCredentials = this.createHasOneRepositoryFactoryFor(
       'userCredentials',
@@ -31,6 +35,10 @@ export class UserRepository extends DefaultCrudRepository<User, typeof User.prot
     this.songs = this.createHasManyRepositoryFactoryFor(
       'songs',
       songRepositoryGetter
+    );
+    this.mediaCredentials = this.createHasManyRepositoryFactoryFor(
+      'mediaCredentials',
+      mediaCredentialsRepositoryGetter
     );
   }
